replaced double with decimal; implemented IHasPrices for all order requests to set price precision; read test order prices from oanda to decouple tests from EURUSD

+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Account/AccountChangesState.cs
@@ -10,19 +10,19 @@ namespace OkonkwoOandaV20.TradeLibrary.DataTypes.Account
    /// </summary>
    public class AccountChangesState
    {
-      public double unrealizedPL { get; set; }
-      public double NAV { get; set; }
-      public double marginUsed { get; set; }
-      public double marginAvailable { get; set; }
-      public double positionValue { get; set; }
-      public double marginCloseoutUnrealizedPL { get; set; }
-      public double marginCloseoutNAV { get; set; }
-      public double marginCloseoutMarginUsed { get; set; }
-      public double marginCloseoutPercent { get; set; }
-      public double marginCloseoutPositionValue { get; set; }
-      public double withdrawalLimit { get; set; }
-      public double marginCallMarginUsed { get; set; }
-      public double marginCallPercent { get; set; }
+      public decimal unrealizedPL { get; set; }
+      public decimal NAV { get; set; }
+      public decimal marginUsed { get; set; }
+      public decimal marginAvailable { get; set; }
+      public decimal positionValue { get; set; }
+      public decimal marginCloseoutUnrealizedPL { get; set; }
+      public decimal marginCloseoutNAV { get; set; }
+      public decimal marginCloseoutMarginUsed { get; set; }
+      public decimal marginCloseoutPercent { get; set; }
+      public decimal marginCloseoutPositionValue { get; set; }
+      public decimal withdrawalLimit { get; set; }
+      public decimal marginCallMarginUsed { get; set; }
+      public decimal marginCallPercent { get; set; }
       public List<DynamicOrderState> orders { get; set; }
       public List<CalculatedTradeState> trades { get; set; }
       public List<CalculatedPositionState> positions { get; set; }
