+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Account/AccountSummary.cs
@@ -8,13 +8,13 @@
       public string id { get; set; }
       public string alias { get; set; }
       public string currency { get; set; }
-      public double balance { get; set; }
+      public decimal balance { get; set; }
       public int createdByUserID { get; set; }
       public string createdTime { get; set; }
-      public double pl { get; set; }
-      public double resettablePL { get; set; }
+      public decimal pl { get; set; }
+      public decimal resettablePL { get; set; }
       public string resettablePLTime { get; set; }
-      public double? marginRate { get; set; }
+      public decimal? marginRate { get; set; }
       public string marginCallEnterTime { get; set; }
       public int marginCallExtensionCount { get; set; }
       public string lastMarginCallExtensionTime { get; set; }
@@ -22,19 +22,19 @@
       public int openPositionCount { get; set; }
       public int pendingOrderCount { get; set; }
       public bool hedgingEnabled { get; set; }
-      public double unrealizedPL { get; set; }
-      public double NAV { get; set; }
-      public double marginUsed { get; set; }
-      public double marginAvailable { get; set; }
-      public double positionValue { get; set; }
-      public double marginCloseoutUnrealizedPL { get; set; }
-      public double marginCloseoutNAV { get; set; }
-      public double marginCloseoutMarginUsed { get; set; }
-      public double marginCloseoutPercent { get; set; }
-      public double marginCloseoutPositionValue { get; set; }
-      public double withdrawalLimit { get; set; }
-      public double marginCallMarginUsed { get; set; }
-      public double marginCallPercent { get; set; }
+      public decimal unrealizedPL { get; set; }
+      public decimal NAV { get; set; }
+      public decimal marginUsed { get; set; }
+      public decimal marginAvailable { get; set; }
+      public decimal positionValue { get; set; }
+      public decimal marginCloseoutUnrealizedPL { get; set; }
+      public decimal marginCloseoutNAV { get; set; }
+      public decimal marginCloseoutMarginUsed { get; set; }
+      public decimal marginCloseoutPercent { get; set; }
+      public decimal marginCloseoutPositionValue { get; set; }
+      public decimal withdrawalLimit { get; set; }
+      public decimal marginCallMarginUsed { get; set; }
+      public decimal marginCallPercent { get; set; }
       public long lastTransactionID { get; set; }
    }
 }
