+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Order/MarketIfTouchedOrder.cs
@@ -7,13 +7,13 @@ namespace OkonkwoOandaV20.TradeLibrary.DataTypes.Order
    {
       public string instrument { get; set; }
       public long units { get; set; }
-      public double price { get; set; }
-      public double? priceBound { get; set; }
+      public decimal price { get; set; }
+      public decimal? priceBound { get; set; }
       public string timeInForce { get; set; }
       public string gtdTime { get; set; }
       public string positionFill { get; set; }
       public string triggerCondition { get; set; }
-      public double initialMarketPrice { get; set; }
+      public decimal initialMarketPrice { get; set; }
       public TakeProfitDetails takeProfitOnFill { get; set; }
       public StopLossDetails stopLossOnFill { get; set; }
       public TrailingStopLossDetails trailingStopLossOnFill { get; set; }
