+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Position/PositionSide.cs
@@ -4,11 +4,11 @@ namespace OkonkwoOandaV20.TradeLibrary.DataTypes.Position
 {
    public class PositionSide
    {
-      public double units { get; set; }
-      public double averagePrice { get; set; }
+      public decimal units { get; set; }
+      public decimal averagePrice { get; set; }
       public List<long> tradeIDs { get; set; }
-      public double pl { get; set; }
-      public double unrealizedPL { get; set; }
-      public double resettablePL { get; set; }
+      public decimal pl { get; set; }
+      public decimal unrealizedPL { get; set; }
+      public decimal resettablePL { get; set; }
    }
 }
