+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Transaction/ClientConfigureRejectTransaction.cs
@@ -3,7 +3,7 @@
    public class ClientConfigureRejectTransaction : Transaction
    {
       public string alias { get; set; }
-      public double marginRate { get; set; }
+      public decimal marginRate { get; set; }
       public string rejectReason { get; set; }
    }
 }
