+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Transaction/ClientConfigureTransaction.cs
@@ -3,6 +3,6 @@
    public class ClientConfigureTransaction : Transaction
    {
       public string alias { get; set; }
-      public double marginRate { get; set; }
+      public decimal marginRate { get; set; }
    }
 }
