+++ OkonkwoOandaV20/OkonkwoOandaV20/TradeLibrary/DataTypes/Transaction/LimitOrderTransaction.cs
@@ -3,8 +3,8 @@
    public class LimitOrderTransaction : Transaction
    {
       public string instrument { get; set; }
-      public double units { get; set; }
-      public double price { get; set; }
+      public decimal units { get; set; }
+      public decimal price { get; set; }
       public string timeInForce { get; set; }
       public string gtdTime { get; set; }
       public string positionFill { get; set; }
